Create separate class for converting firebase data to objects and back. Add unit test for converting Observation back and forth
